Make into a better system
