+++ build-logic/src/main/java/qsl/internal/dependency/QslLibraryDependency.java
@@ -1,21 +1,47 @@
 package qsl.internal.dependency;
 
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
 import org.gradle.api.Named;
 import org.gradle.api.Project;
-import org.gradle.api.internal.AbstractNamedDomainObjectContainer;
-import org.gradle.api.internal.CollectionCallbackActionDecorator;
-import org.gradle.api.model.ObjectFactory;
-import org.gradle.internal.reflect.DirectInstantiator;
+import org.gradle.api.artifacts.Dependency;
+import org.gradle.api.plugins.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 
-public class QslLibraryDependency extends AbstractNamedDomainObjectContainer<QslModuleDependency> implements Named {
+public class QslLibraryDependency  implements Named {
 	private final String name;
-	private final ObjectFactory objects;
+	private final Project project;
+	private final List<Dependency> moduleDependencies;
 
-	public QslLibraryDependency(String name, Project project) {
-		super(QslModuleDependency.class, DirectInstantiator.INSTANCE, Namer.INSTANCE, CollectionCallbackActionDecorator.NOOP);
+	public QslLibraryDependency(String name, Project project, List<Dependency> moduleDependencies) {
 		this.name = name;
-		this.objects = project.getObjects();
+		this.project = project;
+
+		this.moduleDependencies = moduleDependencies;
+	}
+
+	private void add(String module, ConfigurationType type) {
+		Map<String, String> map = new LinkedHashMap<>(2);
+		map.put("path", ":" + name + ":" + module);
+		map.put("configuration", "dev");
+
+		Dependency dep = this.project.getDependencies().project(map);
+		moduleDependencies.add(dep);
+		this.project.getDependencies().add(type.getConfigurationName(), dep);
+	}
+
+	public void api(String module) {
+		add(module, ConfigurationType.API);
+	}
+
+	public void impl(String module) {
+		add(module, ConfigurationType.IMPLEMENTATION);
+	}
+
+	public void testmodOnly(String module) {
+		add(module, ConfigurationType.TESTMOD);
 	}
 
 	@Override
@@ -23,8 +49,19 @@ public class QslLibraryDependency extends AbstractNamedDomainObjectContainer<Qsl
 		return name;
 	}
 
-	@Override
-	protected @NotNull QslModuleDependency doCreate(@NotNull String s) {
-		return objects.newInstance(QslModuleDependency.class, s, objects);
+	/**
+	 * The configuration type for the module dependency
+	 */
+	public enum ConfigurationType {
+		API(JavaPlugin.API_CONFIGURATION_NAME), IMPLEMENTATION(JavaPlugin.IMPLEMENTATION_CONFIGURATION_NAME), TESTMOD("testmodImplementation");
+
+		private final String configurationName;
+		ConfigurationType(String configurationName) {
+			this.configurationName = configurationName;
+		}
+
+		public String getConfigurationName() {
+			return configurationName;
+		}
 	}
 }
