+++ library/command/client_command/build.gradle
@@ -8,7 +8,7 @@ qslModule {
 	library = "command"
 	moduleDependencies {
 		core {
-			qsl_base
+			api("qsl_base")
 		}
 	}
 }
